requirements.txt
absl-py==1.1.0
addict==2.4.0
affine==2.4.0
aiofiles==23.1.0
aiohttp==3.8.5
aiosignal==1.2.0
albumentations==1.2.1
altair==4.2.0
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
argcomplete==3.1.4
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrdict==2.0.1
av==10.0.0
aws-shell==0.2.2
awscli==1.27.73
Babel==2.12.1
backports.zoneinfo==0.2.1
bce-python-sdk==0.8.87
bitarray==2.8.1
bitstring==4.1.2
black==22.6.0
blinker==1.6.2
bokeh==3.1.1
boto3==1.26.73
botocore==1.29.73
cachetools==5.2.0
certifi==2023.7.22
cfn-flip==1.3.0
chromedriver-autoinstaller==0.6.2
chromedriver-binary==2.42.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.2.0
colorama==0.4.4
configobj==5.0.8
contourpy==1.0.7
cssselect==1.2.0
cssutils==2.7.1
cycler==0.11.0
dataclasses-json==0.5.14
dictdiffer==0.9.0
dnspython==2.4.2
docker-pycreds==0.4.0
docutils==0.16
durationpy==0.5
easyocr==1.7.0
efficientnet-pytorch==0.7.1
einops==0.6.1
emoji==2.6.0
et-xmlfile==1.1.0
fairscale==0.4.13
fastapi==0.101.0
ffmpy==0.3.1
fire==0.5.0
Flask==2.3.2
flask-babel==3.1.0
Flask-Cors==4.0.0
flatbuffers==1.12
fonttools==4.41.1
frozenlist==1.3.0
fsspec==2022.5.0
future==0.18.3
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.30
google-api-core==2.12.0
google-api-python-client==2.103.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
googleapis-common-protos==1.61.0
gradio==3.39.0
gradio_client==0.3.0
grpc-gateway-protoc-gen-openapiv2==0.1.0
grpcio==1.59.0
hjson==3.1.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
huggingface-hub==0.16.4
hydra-core==1.3.1
image-similarity-measures==0.3.5
imageio-ffmpeg==0.4.8
imgaug==0.4.0
imutils==0.5.4
iopath==0.1.9
ipython-genutils==0.2.0
itsdangerous==2.1.2
jmespath==1.0.1
joblib==1.1.0
kappa==0.6.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.3
langchain==0.0.254
langsmith==0.0.19
libclang==14.0.1
linkify-it-py==1.0.3
lmdb==1.4.1
loguru==0.7.2
lxml==4.9.3
lz4==4.3.2
marisa-trie==0.7.8
Markdown==3.3.7
markdown-it-py==2.1.0
marshmallow==3.20.1
matplotlib==3.4.3
mdit-py-plugins==0.3.3
mdurl==0.1.2
moviepy==1.0.3
multidict==6.0.2
mutagen==1.46.0
mypy-extensions==0.4.3
ninja==1.11.1
numexpr==2.8.5
numpy==1.23.0
oauth2client==4.1.3
oauthlib==3.2.0
omegaconf==2.3.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
opencv-contrib-python==4.6.0.66
opencv-python==4.5.5
opencv-python-headless==4.6.0.66
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.9.2
osxmetadata==1.3.1
paddle-bfloat==0.1.7
paddleocr==2.6.1.3
paddlepaddle==2.5.1
pandas==1.4.3
parameterized==0.8.1
pathspec==0.9.0
pathtools==0.1.2
pdf2docx==0.5.6
phantomjs==1.4.1
phasepack==1.5
pilmoji==2.0.3
pinecone-client==2.2.4
pipdeptree==2.13.0
placebo==0.9.0
platformdirs==2.5.2
portalocker==2.6.0
premailer==3.10.0
proglog==0.1.10
progressbar2==4.0.0
protobuf==3.20.3
psycopg2==2.9.9
py-applescript==1.0.3
py-spy==0.3.14
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post4
pycocotools==2.0.6
pycryptodome==3.18.0
pycryptodomex==3.18.0
pydantic==1.10.12
pydantic_core==2.4.0
pydeck==0.8.0
pyDeprecate==0.3.2
pydub==0.25.1
pymediainfo==6.0.1
Pympler==1.0.1
PyMuPDF==1.20.2
pyobjc-core==9.2
pyobjc-framework-AppleScriptKit==9.2
pyobjc-framework-AppleScriptObjC==9.2
pyobjc-framework-AVFoundation==9.2
pyobjc-framework-Cocoa==9.2
pyobjc-framework-CoreAudio==9.2
pyobjc-framework-CoreMedia==9.2
pyobjc-framework-CoreServices==9.2
pyobjc-framework-FSEvents==9.2
pyobjc-framework-Quartz==9.2
pyobjc-framework-ScriptingBridge==9.2
PyQt5==5.15.2
PyQt5_sip==4.19.25
PyQtChart==5.15.2
pytesseract==0.3.10
python-bidi==0.4.2
python-docx==0.8.11
python-multipart==0.0.6
python-slugify==8.0.1
python-utils==3.3.3
pytorch-lightning==1.6.5
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
PyYAML==5.4.1
qudida==0.0.4
rapidfuzz==3.1.2
rarfile==4.0
rasterio==1.3.5.post1
requests-oauthlib==1.3.1
rich==13.2.0
rsa==4.7.2
s3transfer==0.6.0
safetensors==0.3.1
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
selenium==4.10.0
semantic-version==2.10.0
semver==2.13.0
sentry-sdk==1.28.1
setproctitle==1.3.2
shapely==2.0.1
simplejson==3.18.1
sklearn==0.0.post1
smmap==5.0.0
snuggs==1.4.7
SQLAlchemy==2.0.19
starlette==0.27.0
streamlit==1.17.0
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.9.0
tensorflow-macos==2.9.2
tensorflow-metal==0.5.0
termcolor==1.1.0
text-unidecode==1.3
thop==0.1.1.post2207130030
threadpoolctl==3.1.0
tifffile==2022.7.28
tiktoken==0.5.1
timm==0.9.5
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchmetrics==0.9.3
torchsummary==1.5.1
torchvision==0.15.2
tqdm==4.66.1
troposphere==4.5.1
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2022.7
tzlocal==4.2
uc-micro-py==1.0.2
uritemplate==4.1.1
uvicorn==0.23.2
validators==0.20.0
visualdl==2.5.3
wandb==0.15.5
webencodings==0.5.1
websockets==11.0.3
Werkzeug==2.3.6
wrapt==1.14.1
xattr==0.10.1
xyzservices==2023.5.0
yacs==0.1.8
yagmail==0.15.293
yarl==1.7.2
yt-dlp==2023.7.6
zappa==0.58.0